+++ app/src/main/java/com/hansung/sherpa/StaticValue.kt
@@ -1,7 +1,7 @@
 package com.hansung.sherpa
 
 import com.hansung.sherpa.itemsetting.TransportRoute
-import com.hansung.sherpa.user.login.User1
+import com.hansung.sherpa.user.`class`.User1
 
 class StaticValue {
     companion object{
