--- app/src/main/java/com/hansung/sherpa/user/createuser/CreateUserResponse.kt
@@ -1,7 +1,0 @@
-package com.hansung.sherpa.user.createuser
-
-data class CreateUserResponse (
-    val code: Int? = null,
-    val message: String? = null,
-    // 추가 필요
-)
