--- app/src/main/java/com/hansung/sherpa/user/linkpermission/LinkPermissionResponse.kt
@@ -1,7 +1,0 @@
-package com.hansung.sherpa.user.linkpermission
-
-class LinkPermissionResponse (
-    val code: Int? = null,
-    val message: String? = null,
-    val data:Int? = null
-)
