--- app/src/main/java/com/hansung/sherpa/user/login/LoginRequest.kt
@@ -1,8 +1,0 @@
-package com.hansung.sherpa.user.login
-
-import com.google.gson.annotations.SerializedName
-
-class LoginRequest (
-    @SerializedName("email") val email: String,
-    @SerializedName("password") val password: String
-)
