--- app/src/main/java/com/hansung/sherpa/user/relation/RelationResponse.kt
@@ -1,7 +1,0 @@
-package com.hansung.sherpa.user.relation
-
-data class RelationResponse(
-    val code: Int? = null,
-    val message: String? = null,
-    val `data`: Relation? = null
-)
